Speed up read tail confirmation
This commit fixes a speed bottleneck of Biopython when it reads the
reference sequence from the genome fasta file. After creating or loading
the SeqIO index, a chromosome is accessed with index["chr1"]. This is
slow (~3-10 sec). Previously, this was done before each and every kmer
counting and even if the same chromosome was read over and over again.
This commit changes the code so that each chromosome is read only once.